fix bug

+++ camus-sweeper/src/main/java/com/linkedin/camus/sweeper/CamusHourlySweeper.java
@@ -233,7 +233,7 @@ public class CamusHourlySweeper extends CamusSweeper {
     public Void call() throws IOException {
       String inputPaths = this.props.getProperty(CamusHourlySweeper.INPUT_PATHS);
       String outputPathStr = this.props.getProperty(CamusHourlySweeper.DEST_PATH);
-      long destinationModTime = getDestinationModTime(outputPathStr);
+      long destinationModTime = getDestinationModTime(this.fs, outputPathStr, true);
       Path outputPath = new Path(outputPathStr, "outlier");
       fs.mkdirs(outputPath);
 
@@ -260,25 +260,30 @@ public class CamusHourlySweeper extends CamusSweeper {
       return null;
     }
 
-    /**
-     * If the destination folder already exist, get the timestamp when the MapReduce job was submitted
-     * to create that folder. The timestamp is stored in a _{timestamp} file.
-     *
-     * If such a file doesn't exist, return the timestamp of the folder.
-     * @throws IOException 
-     */
-    private long getDestinationModTime(String outputPathStr) throws IOException {
-      for (FileStatus status : fs.listStatus(new Path(outputPathStr))) {
-        if (!status.isDir() && status.getPath().getName().matches("_\\d+")) {
-          LOG.info("Found timestamp file: " + status.getPath());
-          long timeStamp = Long.valueOf(status.getPath().getName().substring(1));
-          fs.delete(status.getPath(), false);
-          return timeStamp;
-        }
-      }
+  }
 
-      //return the timestamp of the folder
-      return fs.getFileStatus(new Path(outputPathStr)).getModificationTime();
+  /**
+   * If the destination folder already exist, get the timestamp when the MapReduce job was submitted
+   * to create that folder. The timestamp is stored in a _{timestamp} file.
+   *
+   * If such a file doesn't exist, return the timestamp of the folder.
+   * @throws IOException 
+   */
+  public static long getDestinationModTime(FileSystem fs, String outputPathStr, boolean deleteTimestamp)
+      throws IOException {
+    for (FileStatus status : fs.listStatus(new Path(outputPathStr))) {
+      if (!status.isDir() && status.getPath().getName().matches("_\\d+")) {
+        LOG.info("Found timestamp file: " + status.getPath());
+        long timeStamp = Long.valueOf(status.getPath().getName().substring(1));
+
+        if (deleteTimestamp) {
+          fs.delete(status.getPath(), false);
+        }
+        return timeStamp;
+      }
     }
+
+    //return the timestamp of the folder
+    return fs.getFileStatus(new Path(outputPathStr)).getModificationTime();
   }
 }
